introducing the system component
