+++ SpatialGDK/Source/SpatialGDK/Private/EngineClasses/SpatialNetConnection.cpp
@@ -42,6 +42,7 @@ void USpatialNetConnection::CleanUp()
 {
 	if (USpatialNetDriver* SpatialNetDriver = Cast<USpatialNetDriver>(Driver))
 	{
+		// Clean up should call the client connection manager (Need either a function which calls, or access directly)
 		SpatialNetDriver->CleanUpClientConnection(this);
 	}
