+++ SpatialGDK/Source/SpatialGDK/Private/Interop/ClientConnectionManager.cpp
@@ -1,0 +1,103 @@
+﻿// Copyright (c) Improbable Worlds Ltd, All Rights Reserved
+
+#include "Interop/ClientConnectionManager.h"
+
+#include "EngineClasses/SpatialNetConnection.h"
+#include "EngineClasses/SpatialPackageMapClient.h"
+#include "GameFramework/PlayerController.h"
+#include "Interop/SpatialOSDispatcherInterface.h"
+#include "Interop/SpatialReceiver.h"
+#include "SpatialView/EntityDelta.h"
+#include "SpatialView/SubView.h"
+
+DEFINE_LOG_CATEGORY(LogWorkerEntitySystem);
+
+namespace SpatialGDK
+{
+struct FSubViewDelta;
+
+ClientConnectionManager::ClientConnectionManager(const FSubView& InSubView, USpatialNetDriver* InNetDriver)
+	: SubView(&InSubView)
+	, NetDriver(InNetDriver)
+{
+}
+
+void ClientConnectionManager::Advance()
+{
+	const FSubViewDelta& SubViewDelta = SubView->GetViewDelta();
+	for (const EntityDelta& Delta : SubViewDelta.EntityDeltas)
+	{
+		switch (Delta.Type)
+		{
+		case EntityDelta::REMOVE:
+			EntityRemoved(Delta.EntityId);
+		default:
+			break;
+		}
+	}
+}
+
+void ClientConnectionManager::RegisterClientConnection(const Worker_EntityId InWorkerEntityId, USpatialNetConnection* ClientConnection)
+{
+	WorkerConnections.Add(InWorkerEntityId, ClientConnection);
+}
+
+void ClientConnectionManager::CleanUpClientConnection(USpatialNetConnection* ConnectionCleanedUp)
+{
+	if (ConnectionCleanedUp->ConnectionClientWorkerSystemEntityId != SpatialConstants::INVALID_ENTITY_ID)
+	{
+		WorkerConnections.Remove(ConnectionCleanedUp->ConnectionClientWorkerSystemEntityId);
+	}
+}
+
+void ClientConnectionManager::EntityRemoved(const Worker_EntityId EntityId)
+{
+	// Check to see if we are removing a system entity for a client worker connection. If so clean up the
+	// ClientConnection to delete any and all actors for this connection's controller.
+	if (WorkerConnectionEntities.Contains(EntityId))
+	{
+		const TWeakObjectPtr<USpatialNetConnection> ClientConnectionPtr = FindClientConnectionFromWorkerEntityId(EntityId);
+		if (USpatialNetConnection* ClientConnection = ClientConnectionPtr.Get())
+		{
+			CloseClientConnection(ClientConnection);
+		}
+		WorkerConnectionEntities.Remove(EntityId);
+	}
+}
+
+void ClientConnectionManager::DisconnectPlayer(Worker_EntityId ClientEntityId)
+{
+	Worker_CommandRequest Request = {};
+	Request.component_id = SpatialConstants::WORKER_COMPONENT_ID;
+	Request.command_index = SpatialConstants::WORKER_DISCONNECT_COMMAND_ID;
+	Request.schema_type = Schema_CreateCommandRequest();
+	Worker_RequestId RequestId = NetDriver->Connection->SendCommandRequest(ClientEntityId, &Request, RETRY_UNTIL_COMPLETE, {});
+
+	SystemEntityCommandDelegate CommandResponseDelegate;
+	CommandResponseDelegate.BindWeakLambda(this, [this, ClientEntityId](const Worker_CommandResponseOp& Op) {
+		TWeakObjectPtr<USpatialNetConnection> ClientConnection = FindClientConnectionFromWorkerEntityId(ClientEntityId);
+		if (ClientConnection.IsValid())
+		{
+			ClientConnection->CleanUp();
+		}
+	});
+
+	NetDriver->Receiver->AddSystemEntityCommandDelegate(RequestId, CommandResponseDelegate);
+}
+
+TWeakObjectPtr<USpatialNetConnection> ClientConnectionManager::FindClientConnectionFromWorkerEntityId(const Worker_EntityId WorkerEntityId)
+{
+	if (TWeakObjectPtr<USpatialNetConnection>* ClientConnectionPtr = WorkerConnections.Find(WorkerEntityId))
+	{
+		return *ClientConnectionPtr;
+	}
+
+	return {};
+}
+
+void ClientConnectionManager::CloseClientConnection(USpatialNetConnection* ClientConnection)
+{
+	ClientConnection->CleanUp();
+}
+
+} // Namespace SpatialGDK
